+++ test/definition.test.ts
@@ -12,12 +12,10 @@ import {
   collect,
   createPiAgent,
   createPiAgentFromDefinition,
-  loadAgentDefinition,
-  piBasePrompt,
-  piDefaultTools,
   type CreatePiAgentFromDefinitionOptions,
 } from "../src/index.ts";
-import { assembleSystemPrompt } from "../src/engines/pi/create.ts";
+import { assembleSystemPrompt, piBasePrompt, piDefaultTools } from "../src/engines/pi/create.ts";
+import { loadAgentDefinition } from "../src/engines/pi/definition.ts";
 import { log } from "../src/log.ts";
 import { ensureStateRootSelfIgnored, isUnderDir } from "../src/engines/pi/definition.ts";